requirements.txt
streamlit
numpy
opencv-python
keras
matplotlib
pillow
tensorflow
dlib
seaborn
numpy